% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findWords1.R
\name{findWords1}
\alias{findWords1}
\title{list the words in a text file}
\usage{
findWords1(pfn = NULL)
}
\arguments{
\item{pfn}{- a text file possibly with a path included.}
}
\value{
A list of the words and their sequence number in the file
}
\description{
list the words in a text file
}

