% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findWords2.R
\name{findWords2}
\alias{findWords2}
\title{findwords2: elegant improvement on findwords1}
\usage{
findWords2(pfn = NULL)
}
\arguments{
\item{txt}{- an .Rda file containing a text file}
}
\value{
A list of the words and their sequence number in the file
}
\description{
findwords2: elegant improvement on findwords1
}
\examples{
findwords2()
}

